+++ GN Danavox 2026 Checker.py
@@ -52,7 +52,7 @@ if not turboFile.is_file():
 # Get Danavox XE BeMore update from the webpage
 dxebmURI = rot_codec.rot47_decode("9EEADi^^HHH]8?962C:?8]4@>^6?^AC@5F4ED^52?2G@I^") + "fitting-software-download"
 try:
-    test = requests.get(dxebmURI)
+    test = requests.get(dxebmURI, verify='www-gnhearing-com-chain.pem')
     dom = lxml.html.fromstring(test.content)
     hrefs = [x for x in dom.xpath('//a/@href') if '//' in x and 'zip' in x]
     link0 = hrefs[0].replace('%20', ' ')