+++ GN Interton 2026 Checker.py
@@ -58,7 +58,7 @@ if not turboFile.is_file():
 # Get Interton Fitting update from the webpage
 ifURI = rot_codec.rot47_decode("9EEADi^^HHH]8?962C:?8]4@>^6?^AC@5F4ED^:?E6CE@?^") + "fitting-software-download"
 try:
-    test = requests.get(ifURI)
+    test = requests.get(ifURI, verify='www-gnhearing-com-chain.pem')
     dom = lxml.html.fromstring(test.content)
     hrefs = [x for x in dom.xpath('//a/@href') if '//' in x and 'zip' in x]
     link0 = hrefs[0].replace('%20', ' ')